backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
schwab-py==1.4.0
yfinance==0.2.43
python-dotenv==1.0.1
